# Update nlohmann/json library (fixes #284)