--- a/resources/3rdparty/modernjson/CITATION.cff
+++ b/resources/3rdparty/modernjson/CITATION.cff
@@ -1,0 +1,14 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: Lohmann
+    given-names: Niels
+    orcid: https://orcid.org/0000-0001-9037-795X
+    email: mail@nlohmann.me
+    website: https://nlohmann.me
+title: "JSON for Modern C++"
+version: 3.11.2
+date-released: 2022-08-12
+license: MIT
+repository-code: "https://github.com/nlohmann"
+url: https://json.nlohmann.me