--- a/resources/3rdparty/modernjson/LICENSE.MIT
+++ b/resources/3rdparty/modernjson/LICENSE.MIT
@@ -1,14 +1,13 @@
-The library is licensed under the MIT License 
-<http://opensource.org/licenses/MIT>:
+MIT License 
 
-Copyright (c) 2013-2016 Niels Lohmann
+Copyright (c) 2013-2022 Niels Lohmann
 
-Permission is hereby granted, free of charge, to any person obtaining a copy of
-this software and associated documentation files (the "Software"), to deal in
-the Software without restriction, including without limitation the rights to
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
-of the Software, and to permit persons to whom the Software is furnished to do
-so, subject to the following conditions:
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
 
 The above copyright notice and this permission notice shall be included in all
 copies or substantial portions of the Software.