--- a/resources/3rdparty/modernjson/WORKSPACE.bazel
+++ b/resources/3rdparty/modernjson/WORKSPACE.bazel
@@ -1,0 +1,1 @@
+workspace(name = "nlohmann_json")