I re-applied the [changes](https://github.com/moves-rwth/storm/commit/632c9c2e1e7cdfe36aba1287788021261671164f#diff-e1d25a1c865ce280fc884d097b0030464e4f3e58d24acc4baa0981d3d69a9671) made to the old json version concerning RationalNumber support.

My changes to the main nlohmann/json release can be seen [here](https://github.com/tquatmann/json/commit/48ebe445314e798af208851374fb9508a45d199a)

I briefly checked for obvious performance regressions, which does not seem to be the case.
Here some data, where I used storm-conv to parse and then export a ~38 MB JANI file from QVBS:

current main:
```console
Command line arguments: --jani oscillators.8-10-0.1-1.jani --tojani test1 -jprop --compactjson
Current working directory: /Users/tim/git/qcomp/benchmarks/dtmc/oscillators

Parsing JANI input ...  done. (2.175s seconds).
Performing transformations on JANI model ...  done. (0.201s seconds).
Exporting JANI model ... Stored to file 'test1' done. (1.268s seconds).
```

This branch:
```console
Command line arguments: --jani oscillators.8-10-0.1-1.jani --tojani test1 -jprop --compactjson
Current working directory: /Users/tim/git/qcomp/benchmarks/dtmc/oscillators

Parsing JANI input ...  done. (2.199s seconds).
Performing transformations on JANI model ...  done. (0.200s seconds).
Exporting JANI model ... Stored to file 'test1' done. (1.081s seconds).
?master ~/git/qcomp/benchmarks/dtmc/oscillators> 
```