@tquatmann 2023-11-10
I still need to add some documentation [here](https://github.com/moves-rwth/storm/blob/master/doc/update_resources.md) before merging this.