@sjunges 2023-11-11
Is there any reason why we did not specialize std::is_scalar_v in the changes within modernjson? That would have saved quite some lines in the diff?