@tquatmann 2023-11-11
specialising is_scalar (and similar traits) would be great, but (quoting [cppreference](https://en.cppreference.com/w/cpp/types/is_scalar) ):

> The behavior of a program that adds specializations for std::is_scalar or std::is_scalar_v is undefined.