@tquatmann 2023-11-14
I made the json export a bit more robust (in case a huge `storm::RationalNumber` like `1E5000` might be converted to something like `nan` and/or `inf`.

I also added the promised documentation. From my side, this is ready to be merged :)