@marisgg 2023-11-21
The Stormpy master fork fails (https://github.com/randriu/stormpy/actions/runs/6945132931/job/18893799741) on the latest version of Storm CI, possibly due to this PR. The error is inside `CMakeFiles/utility.dir/src/utility/json.cpp`.

I'm assuming this very recently started Stormpy build (https://github.com/moves-rwth/stormpy/actions/runs/6945494052/job/18894977397) will fail because of this.