@sjunges 2023-11-21
Looking at this, the problem is here: https://github.com/randriu/stormpy/blob/master/src/utility/json.cpp but I dont know where exactly. Maybe @tquatmann sees it?